+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for ui-checkin
 
+* Add support for use circulation-bff for inventory-items. Refs UICHKIN-488.
+
 ## [11.0.1] (https://github.com/folio-org/ui-checkin/tree/v11.0.1) (2025-04-29)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v11.0.0...v11.0.1)